power-libperfmgr: PowerHAL: fix deadlock while PowerHintSession calls to DoHint
The root cause of the deadlock is:
1. When NodeLooperThread thread EventNode updated AdpfProfile for
   a session, it held `NodeLooperThread.lock_` and tried to take
   `mPowerHintSessionLock` to update the Session.mAdpfProfile.
2. At the same time, the ADPF binder call sendHint(CPU_LOAD_RESET)
   held the `mPowerHintSessionLock` first and performed DoHint which
   also requires `NodeLooperThread.lock_` to do `Request()`.

To achieve ADPF tag profile switch mechanism, holding two locks in
path-1 is inevitable.
In path-2, we don't need to hold the `mPowerHintSessionLock` while
DoHint since it doesn't operate the Session data.
So the solution is move the DoHint operation out of the
`mPowerHintSessionLock` scope.

Bug: 365906700
Test: atest libperfmgr_test
Test: Manual App cycling test

Change-Id: I833eff636029a55d684634c7771723df03bf6f8b
Signed-off-by: HeroBuxx <me@herobuxx.me>